+++ src/Services/Dominion/Actions/EspionageActionService.php
@@ -239,7 +239,7 @@ class EspionageActionService
                     $bountyRewardString = '';
                     $rewards = $result['bounty'];
                     if (isset($rewards['xp']) && $rewards['xp']) {
-                        $xpGain = max($xpGain, $rewards['xp']);
+                        $xpGain += $rewards['xp'];
                     }
                     if (isset($rewards['resource']) && isset($rewards['amount'])) {
                         $dominion->{$rewards['resource']} += $rewards['amount'];